@@ -1,4 +1,4 @@
-class CreateSongs < ActiveRecord::Migration
+class CreateSongs < ActiveRecord::Migration[4.2]
   def change
     create_table :songs do |t|
       t.string :title
