@@ -1,4 +1,4 @@
-class AddArtistToSong < ActiveRecord::Migration
+class AddArtistToSong < ActiveRecord::Migration[4.2]
   def change
     add_reference :songs, :artist, index: true, foreign_key: true
   end
