@@ -24,9 +24,8 @@ class SongsController < ApplicationController
   end
 
   def upload
-    CSV.foreach(params["file"].path, headers: true) do |song|
-      Song.create(title: song[0], artist_name: song[1])
-    end
+    binding.pry
+    SongsWorker.perform_async(params[:file].path)
     redirect_to songs_path
   end
 
@@ -59,4 +58,3 @@ class SongsController < ApplicationController
     params.require(:song).permit(:title, :artist_name)
   end
 end
-
